+++ package.json
@@ -23,6 +23,7 @@
         "react-paginate": "^8.3.0",
         "react-router-dom": "^7.11.0",
         "react-spinners": "^0.17.0",
+        "tailwind-scrollbar-hide": "^4.0.0",
         "tailwindcss": "^4.1.18",
         "uuid": "^13.0.0",
         "zustand": "^5.0.10"
