+++ src/pages/boards/recipe/recipe-edit-page/RecipeEditPage.jsx
@@ -1,16 +1,8 @@
-import { useNavigate, useParams } from 'react-router-dom';
-import { RecipeEdit } from '../../../../components/RecipeEdit';
+import { useParams } from 'react-router-dom';
+import { RecipeEdit } from '../../../../components/Recipe/RecipeEdit';
 
 export default function RecipeEditPage() {
     const { boardId, recipeId } = useParams();
-    const navigate = useNavigate();
 
-    const onNavigate = (pageKey) => {
-        if (pageKey === 'board') navigate(`/boards/${boardId}/recipe`);
-        if (pageKey === 'home') navigate(`/`);
-        if (pageKey === 'recipe')
-            navigate(`/boards/${boardId}/recipe/${recipeId}`);
-    };
-
-    return <RecipeEdit onNavigate={onNavigate} recipeId={Number(recipeId)} />;
+    return <RecipeEdit recipeId={Number(recipeId)} boardId={Number(boardId)} />;
 }
