+++ vite.config.js
@@ -54,6 +54,11 @@ export default defineConfig({
                 target: 'http://localhost:8080',
                 changeOrigin: true,
             },
+
+            '/recipes/': {
+                target: 'http://localhost:8080',
+                changeOrigin: true,
+            },
         },
     },
 });
